BACKLOG-21943: Split pageBuilder tests in separate suite
