+++ .github/workflows/on-code-change.yml
@@ -81,7 +81,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        config_file: ["cypress.config-contentEditor.ts", "cypress.config-jcontent.ts", "cypress.config-menuActions.ts", "cypress.config-pickers.ts"]
+        config_file: ["cypress.config-contentEditor.ts", "cypress.config-jcontent.ts", "cypress.config-menuActions.ts", "cypress.config-pageBuilder.ts", "cypress.config-pickers.ts"]
     timeout-minutes: 120
     steps:
       - uses: jahia/jahia-modules-action/helper@v2
