+++ .github/workflows/on-merge.yml
@@ -60,7 +60,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        config_file: [ "cypress.config-contentEditor.ts", "cypress.config-jcontent.ts", "cypress.config-menuActions.ts", "cypress.config-pickers.ts" ]
+        config_file: [ "cypress.config-contentEditor.ts", "cypress.config-jcontent.ts", "cypress.config-menuActions.ts", "cypress.config-pageBuilder.ts", "cypress.config-pickers.ts" ]
     steps:
       - uses: jahia/jahia-modules-action/helper@v2
       - uses: KengoTODA/actions-setup-docker-compose@main
