+++ tests/cypress.config.common.ts
@@ -16,6 +16,8 @@ export const baseConfig = {
     viewportWidth: 1366,
     viewportHeight: 768,
     watchForFileChanges: false,
+    experimentalMemoryManagement: true,
+    numTestsKeptInMemory: 20,
     e2e: {
         specPattern: ['**/**.cy.ts'],
         // We've imported your old cypress plugins here.
